+++ src/DB/Data/Connection.php
@@ -78,12 +78,4 @@ class Connection implements ConnectionInterface
     {
         return $this->connectionData['password'] ?? '';
     }
-
-    /**
-     * @inheritdoc
-     */
-    public function getTablePrefix(): string
-    {
-        return $this->connectionData['table_prefix'] ?? '';
-    }
 }
